% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_put_payoff.R
\name{binary_put_payoff}
\alias{binary_put_payoff}
\title{Calculate european binary put option payoff}
\usage{
binary_put_payoff(asset_price, strike)
}
\arguments{
\item{asset_price}{numeric vector, price of the asset over time.}

\item{strike}{numeric value, strike price of the option.}
}
\value{
A numeric value, european binary put option payoff.
}
\description{
The binary_put_payoff function calculate payoffs from trajectory.
}
\examples{
x <- generate_scenarios(100, 0, 0.3)
binary_put_payoff(x[, 1], strike = 100)
binary_put_payoff(c(100, 120), strike = 107)
}
