% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bisection_method.R
\name{Bisection_method}
\alias{Bisection_method}
\title{Bisection method}
\usage{
Bisection_method(option_value, a0, b0, FUN, ..., epsilon = 1e-7)
}
\arguments{
\item{option_value}{numeric value, price of the modified option.}

\item{a0}{numeric value, lower boundary of the range.}

\item{b0}{numeric value, upper boundary of the range.}

\item{FUN}{the function to be applied, explicit formula for the option price.}

\item{...}{arguments to FUN.}

\item{epsilon}{calculation error.}
}
\value{
A numeric value, a constant for which the price of the modified option is equal to option_value.
}
\description{
The Bisection_method function is used to find const L in modified call and put option.
}
\examples{
Bisection_method(5, 100, 500, call_price_linear, asset = 100, strike = 100, rate = 0, vol = 0.3, drift = 0.1, time = 0, End_Time = 1)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Bisection_method}
}
