% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bisection_method_MC.R
\name{Bisection_method_MC}
\alias{Bisection_method_MC}
\title{Bisection method for Monte Carlo simulations}
\usage{
Bisection_method_MC(option_value, a0, b0, rate, X, FUN, ..., epsilon = 1e-3, time = 0, End_Time = 1)
}
\arguments{
\item{option_value}{numeric value, price of the modified option.}

\item{a0}{numeric value, lower boundary of the range.}

\item{b0}{numeric value, upper boundary of the range.}

\item{rate}{numeric value, risk free rate in the model rate >= 0.}

\item{X}{numeric matrix, price of the asset from different trajectories.}

\item{FUN}{the function to be applied, explicit formula for the option price.}

\item{...}{arguments to FUN.}

\item{epsilon}{calculation error.}

\item{time}{numeric value, actual time.}

\item{End_Time}{numeric value, end time of the option, time < End_Time.}
}
\value{
A numeric value, a constant for which the price of the modified option is equal to option_value.
}
\description{
The Bisection_method_MC function is used to find const L in modified european option.
}
\details{
This function estimated a constant for which price of the modified option is equal to option_value. If warning occurs means that
epsilon parameter is too small to find constant for which calculation error of option price is smaller then epsilon.
The reason for this is simulations. When the number of trajectories is increased, the epsilon parameter can be decreased.
}
\examples{
x <- generate_scenarios(100, 0, 0.3)
Bisection_method_MC(5, 1, 1000, 0, x, option_modificate_payoff, drift = 0.05, vol = 0.3, p = 1, call_payoff, strike = 100)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Bisection_method}
}
