% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_price_linear.R
\name{call_price_linear}
\alias{call_price_linear}
\title{Calculate modified european call option}
\usage{
call_price_linear(asset, strike, rate, vol, drift, time, End_Time, L)
}
\arguments{
\item{asset}{a numeric vector of asset prices.}

\item{strike}{numeric value, strike price for call or put option.}

\item{rate}{numeric value, risk free rate in the model, r >= 0.}

\item{vol}{numeric value, volatility of the model, vol > 0.}

\item{drift}{numeric value, drift of the model.}

\item{time}{a numeric vector of actual time, time > 0.}

\item{End_Time}{end time of the option, End_time >= time.}

\item{L}{a numeric value, determines where option payoff is zero, see details, L > 0.}
}
\value{
A numeric vector, price of the modification of european call option using linear loss function.
}
\description{
The call_price_linear function takes parameters from Black-Scholes model and returns a price of modified european call option.
}
\details{
Payoff of this modified call option is:
\subsection{\eqn{ 1(asset > L)(asset - strike)^+ }, when \eqn{ drift > rate }.}{
}

\subsection{\eqn{ 1(asset < L)(asset - strike)^+ }, when \eqn{ drift < rate }.}{
}

\subsection{\eqn{ L(asset - strike)^+ }, when \eqn{ drift == rate }, of course in this case L <= 1.}{
}
}
\examples{
call_price_linear(100, 100, 0, 0.5, 0.05, 0, 1, 120)
call_price_linear(c(100, 120), 100, 0, 0.3, 0.05, 0, 1, 120)
call_price_linear(c(100, 120), 100, 0, 0.3, 0.05, c(0, 0.5), 1, 120)



}
