% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d1.R
\name{d1}
\alias{d1}
\title{d1}
\usage{
d1(asset, strike, rate, vol, time, End_Time)
}
\arguments{
\item{asset}{a numeric vector of asset prices.}

\item{strike}{numeric value, strike price for call or put option.}

\item{rate}{numeric value, risk free rate in the model, r >= 0.}

\item{vol}{numeric value, volatility of the model, vol > 0.}

\item{time}{a numeric vector of actual time, time > 0.}

\item{End_Time}{end time of the option, End_time >= time.}
}
\value{
A numeric vector, which help calculate option value.
}
\description{
The d1 function is used in an analytical approach to calculate standard call and put options and their modifications.
}
\examples{
d1(100, 100, 0, 0.5, 0, 1)
d1(c(100, 120), 100, 0, 0.3, 0, 1)
d1(c(100, 120), 100, 0, 0.3, c(0, 0.5), 1)



}
\seealso{
\url{https://en.wikipedia.org/wiki/Black–Scholes_model}.

}
