% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_finite_difference.R
\name{delta_finite_difference}
\alias{delta_finite_difference}
\title{Hedging of the option}
\usage{
delta_finite_difference(option_value, ds, is_continuous = TRUE, discontinuity_points = NA)
}
\arguments{
\item{option_value}{numeric matrix, option prices, return from finite_difference_explicit function.}

\item{ds}{numeric value, asset step size.}

\item{is_continuous}{logical value, TRUE means that the payoff function is continuous due to the asset price.}

\item{discontinuity_points}{numeric vector, denoting at which points of asset price the payoff function is discontinuous.
Elements of this vector have to be positive.}
}
\value{
A numeric matrix, estimated number of asset needed to hedge an option.
}
\description{
The delta_finite_difference function calculate number of asset needed to hedge an option using finite difference.
}
\examples{
ds <- finding_parameters(100, 0.3, 600, 5)[1]
dt <- finding_parameters(100, 0.3, 600, 5)[2]
option <- finite_difference_explicit(ds, dt, 600, 0, 0.3, 1, call_payoff, 100)
delta_finite_difference(option, ds)
option <- finite_difference_explicit(ds, dt, 600, 0, 0.3, 1, FUN = option_modificate_payoff, const = 120, drift = 0.1, vol = 0.3, p = 1, call_payoff, 100, is_modificate = TRUE)
delta_finite_difference(option, ds)
}
