% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eta.R
\name{Eta}
\alias{Eta}
\title{Delta hedging for european option}
\usage{
Eta(asset, Xi, rate, time)
}
\arguments{
\item{asset}{a numeric vector of asset prices.}

\item{Xi}{a numeric vector, number of asset needed to hedge an option}

\item{rate}{numeric value, risk free rate in the model, r >= 0.}

\item{option_price}{a numeric vector of option prices}

\item{time}{a numeric vector of actual time, time >= 0.}
}
\value{
A numeric vector, amount of money needed to hedge an option.
}
\description{
The Eta function takes parameters from Black-Scholes model and returns amount of money needed to hedge an european option.
}
\examples{
Eta(100, 0.5596177, 0, 11.92354, 0)
Eta(c(100, 120), c(0.5596177, 0.7756962), 0, c(11.92354, 25.44056), 0)
Eta(c(100, 120), c(0.5596177, 0.8328623), 0, c(11.92354, 22.50378), c(0, 0.5))



}
\seealso{
\url{https://en.wikipedia.org/wiki/Black–Scholes_model}.

}
