% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/False_position_method.R
\name{False_position_method}
\alias{False_position_method}
\title{False position method}
\usage{
False_position_method(V0, a0, b0, FUN, ..., epsilon = 1e-7)
}
\arguments{
\item{option_value}{numeric value, price of the modified option.}

\item{a0}{numeric value, lower boundary of the range.}

\item{b0}{numeric value, upper boundary of the range.}

\item{FUN}{the function to be applied, explicit formula for the option price.}

\item{...}{arguments to FUN.}

\item{epsilon}{calculation error.}
}
\value{
A numeric value, a constant for which the price of the modified option is equal to option_value.
}
\description{
The False_position_method function is used to find const L in modified call and put option.
}
\examples{
False_position_method(5, 100, 500, call_price_linear, asset = 100, strike = 100, rate = 0, vol = 0.3, drift = 0.1, time = 0, End_Time = 1)

}
\seealso{
\url{https://en.wikipedia.org/wiki/Regula_falsi}
}
