% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finding_parameters.R
\name{finding_parameters}
\alias{finding_parameters}
\title{Finite difference parameters}
\usage{
finding_parameters(initial_price, vol, I, multiplier_price, dt_simulate = 1/250)
}
\arguments{
\item{initial_price}{numeric value, initial price of the asset.}

\item{vol}{numeric value, volatility of the model, vol > 0.}

\item{I}{numeric value, number of steps in the stock price.}

\item{multiplier_price}{numeric value, a number multiplied by the initial_price to get an approximation to infinity.}

\item{dt_simulate}{numeric value, time step size using in simulations.}
}
\value{
A numeric vector, respectively price step size and time step size.
}
\description{
The finding_parameters function calculate calculate time and price step sizes.
}
\details{
Explicit finite difference algorithm is not stable, it means we have an limitation on size of asset step size (which is not problematic)
and on size of time step size (which is problematic). This function calculate good parameters to the asset and step size for stability of explicit
finite difference algorithm. \cr \cr
dt_simulate argument is needed for a sake of interpolation. We want existing a natural k which meets:
\eqn{ dt = dt_simulate/k }, where dt is step size from finite difference algorithm. If dt is very small is no needed to have this property.
}
\examples{
finding_parameters(100, 0.3, 600, 5)
finding_parameters(100, 0.5, 1000, 6, 1/500)

}
