% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_interpolation.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{Linear interpolation}
\usage{
linear_interpolation(x, x1, x2, y1, y2)
}
\arguments{
\item{x}{numeric value, argument in the searched function value.}

\item{x1}{numeric value, argument of the function x1 <= x.}

\item{x2}{numeric value, actual argument of the function x <= x2.}

\item{y1}{numeric value, the value of the function at the point x1.}

\item{y2}{numeric value, the value of the function at the point x2.}
}
\value{
A numeric value, interpolated value of the function at the point x.
}
\description{
The linear interpolation function, interpolate value of the function at point x.
}
\examples{
linear_interpolation(3, 2, 3.5, 5, 7)

}
