% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losses.R
\name{losses}
\alias{losses}
\title{Calculate gains/losses}
\usage{
losses(option_value, portfolio)
}
\arguments{
\item{option_value}{numeric vector, price of the option over time.}

\item{portfolio}{numeric vector, value of the portfolio over time.}
}
\value{
A numeric vector, gains/losses during hedging.
}
\description{
The losses function calculate gains/losses between portfolio and option price.
}
\examples{
x <- generate_scenarios(100, 0.05, 0.3) #real measure
time_period <- seq(0, 1, 1/250)
option <- call_price(x[, 1], 100, 0, 0.3, time_period, 1)
Xi <- Xi_call_price(x[, 1], 100, 0, 0.3, time_period, 1)
portfolio <- portfolio_valuation(option, 0, Xi, x[, 1])
losses(option, portfolio)
}
