% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_eta.R
\name{matrix_eta}
\alias{matrix_eta}
\title{Hedging of the option}
\usage{
matrix_eta(option_matrix, Xi_matrix, ds, dt, I, rate)
}
\arguments{
\item{option_matrix}{numeric matrix, value of the option.}

\item{Xi_matrix}{numeric matrix, number of asset needed to hedge an option.}

\item{ds}{numeric value, step size of the asset price.}

\item{dt}{numeric value, step size of the time.}

\item{I}{numeric value, number of steps in the stock price.}

\item{rate}{numeric value, risk free rate in the model rate >= 0.}
}
\value{
A numeric matrix, amount of money needed to hedge an option claim using finite difference algorithms.
}
\description{
The matrix_eta function calculate amount of money needed to hedge an option claim.
}
\examples{
ds <- finding_parameters(100, 0.3, 600, 5)[1]
dt <- finding_parameters(100, 0.3, 600, 5)[2]
option <- finite_difference_explicit(ds, dt, 600, 0, 0.3, 1, call_payoff, 100)
Xi <- delta_finite_difference(option, ds)
matrix_eta(option, Xi, ds, dt, 600, 0)

}
