% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_interpolation.R
\name{matrix_interpolation}
\alias{matrix_interpolation}
\title{Linear interpolation}
\usage{
matrix_interpolation(Value_matrix, ds, dt, I, asset_price, time, End_Time)
}
\arguments{
\item{Value_matrix}{numeric matrix, returns from finite different algorithm or delta_finite_difference function.}

\item{dt}{numeric value, step size of the time.}

\item{ds}{numeric value, step size of the asset price.}

\item{I}{numeric value, number of steps in the stock price.}

\item{asset_price}{numeric vector, the price of the asset over the life of the option.}

\item{time}{numeric vector, actual time.}

\item{End_Time}{numeric value, end time of the option.}
}
\value{
A numeric vector, interpolated value of option price or numbers of asset when asset_price is an argument.
}
\description{
The matrix_interpolation function interpolate values for vector.
}
\details{
If asset price is bigger then I\emph{ds, we take value from this element of the matrix where asset_price = I}ds.
}
\examples{
ds <- finding_parameters(100, 0.3, 600, 5)[1]
dt <- finding_parameters(100, 0.3, 600, 5)[2]
option <- finite_difference_explicit(ds, dt, 600, 0, 0.3, 1, call_payoff, 100)
matrix_interpolation(option, ds, dt, 600, c(259, 150), c(0, 0.5), 1)

}
