% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_linear_payoff.R
\name{option_linear_payoff}
\alias{option_linear_payoff}
\title{Calculate payoff from modificate european option}
\usage{
option_linear_payoff(const, asset_price, drift, rate, FUN, ...)
}
\arguments{
\item{const}{numeric value, parameter of the modified payoff.}

\item{asset_price}{numeric vector, the price of the asset over the life of the option.}

\item{drift}{numeric value, drift of the model.}

\item{rate}{numeric value, risk free rate in the model rate >= 0.}

\item{FUN}{the function to be applied, payoff option: see 'Details' and 'Examples'.}

\item{...}{arguments to FUN}
}
\value{
A numeric value, payoff of the modified european option using linear loss function.
}
\description{
The option_linear_payoff function is used to calculate payoff of modified european option.
}
\details{
FUN have to had parameter asset_price which represent price oft the asset over the life of the option, see call_payoff \cr \cr
Arguments in ... cannot have the same name as any of the other arguments, and care may be needed to avoid errors.
}
\examples{
option_linear_payoff(110, c(121.1, 120, 125.4), 0.1, 0.05, call_payoff, strike = 105)
option_linear_payoff(110, c(141.1, 150, 135.4), 0.1, 0.05, put_payoff, strike = 150)
## Example with asian call payoff
asian_call <- function(asset_price, strike){
  return(mean(asset_price) - min( mean(asset_price), strike ))
  }
option_linear_payoff(110, c(121.1, 120, 125.4), 0.1, 0.05, asian_call, strike = 105)

}
