% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_finding_linear.R
\name{put_finding_linear}
\alias{put_finding_linear}
\title{Finding constant in modified option.}
\usage{
put_finding_linear(asset, strike, rate, vol, drift, time, End_Time, option_value, FUN, ...)
}
\arguments{
\item{asset}{a numeric value of asset prices.}

\item{strike}{numeric value, strike price for put option.}

\item{rate}{numeric value, risk free rate in the model, r >= 0.}

\item{vol}{numeric value, volatility of the model, vol > 0.}

\item{drift}{numeric value, drift of the model.}

\item{time}{a numeric value of actual time, time > 0.}

\item{End_Time}{end time of the option, End_time >= time.}

\item{option_value}{numeric value, price of the modified option.}

\item{FUN}{the function to be applied, numeric method to find constant.}

\item{...}{arguments to FUN.}
}
\value{
A numeric value, a constant for which the price of the modified option is equal to option_value.
}
\description{
The put_finding_linear function is used to find constant L in modified put option.
}
\examples{
put_finding_linear(100, 100, 0, 0.3, 0.1, 0, 1, 2, Bisection_method)

}
