% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_finding_modified.R
\name{put_finding_modified}
\alias{put_finding_modified}
\title{Finding constant in modified option.}
\usage{
put_finding_modified(asset, strike, rate, vol, drift, p, time, End_Time, option_value, FUN, ...)
}
\arguments{
\item{asset}{a numeric value of asset prices.}

\item{strike}{numeric value, strike price for put option.}

\item{rate}{numeric value, risk free rate in the model, r >= 0.}

\item{vol}{numeric value, volatility of the model, vol > 0.}

\item{drift}{numeric value, drift of the model.}

\item{p}{numeric value, power of the loss function, p > 0.}

\item{time}{a numeric value of actual time, time > 0.}

\item{End_Time}{end time of the option, End_time >= time.}

\item{option_value}{numeric value, price of the modified option.}

\item{FUN}{the function to be applied, numeric method to find constant.}

\item{...}{arguments to FUN.}
}
\value{
A numeric vector, a constants for which the price of the modified option is equal to option_value.
}
\description{
The put_finding_modified function is used to find constant L in modified put option.
}
\details{
If an error occurs: "Too large option_value or wrong initial values. Select other parameters." occur, try to use the function
Bisection_method function with different a0 and b0 parameter.
}
\examples{
put_finding_modified(100, 100, 0, 0.3, 0.1, 2, 0, 1, 2, Bisection_method)
put_finding_modified(100, 100, 0, 0.3, 0.1, 1, 0, 1, 2.5, Bisection_method)

}
