% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_Newton_concave.R
\name{put_Newton_concave}
\alias{put_Newton_concave}
\title{Finding constant in modified option.}
\usage{
put_Newton_concave(L1, strike, drift, rate, vol, p, epsilon = 1e-10)
}
\arguments{
\item{L1}{numeric value, first constant L1 > K.}

\item{strike}{numeric value, strike price for put option.}

\item{drift}{numeric value, drift of the model.}

\item{rate}{numeric value, risk free rate in the model, r >= 0.}

\item{vol}{numeric value, volatility of the model, vol > 0.}

\item{p}{numeric value, power of the loss function, p > 1.}

\item{epsilon}{numeric value, acceptable calculation error}
}
\value{
A numeric value, a second constant for modified put option by concave function.
}
\description{
The put_Newton_concave function is used to find second constant L in modified put option.
}
\details{
There is a need to be careful when there is a warning message.
It means that a numerical error occurred during the algorithm and the algorithm was terminated without a while loop
}
\examples{
put_Newton_concave(70, 100, 0, 0.1, 0.2, 0.5)

}
