% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_payoff.R
\name{put_payoff}
\alias{put_payoff}
\title{Calculate european put option payoff}
\usage{
put_payoff(asset_price, strike)
}
\arguments{
\item{asset_price}{numeric vector, price of the asset over time.}

\item{strike}{numeric value, strike price of the option.}
}
\value{
A numeric value, european put option payoff.
}
\description{
The put_payoff function calculate payoffs from trajectory.
}
\examples{
x <- generate_scenarios(100, 0, 0.3)
put_payoff(x[, 1], strike = 100)
put_payoff(c(100, 120), strike = 107)
put_payoff(c(100, 120, 107), strike = 107)
}
