% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shortfall_risk.R
\name{Shortfall_risk}
\alias{Shortfall_risk}
\title{Calculate Shortfall risk}
\usage{
Shortfall_risk(payoff, modificate_payoff, p, rate, End_Time = 1, is_discounted = FALSE)
}
\arguments{
\item{payoff}{a numeric vector of option payoff from trajectory simulations.}

\item{modificate_payoff}{a numeric vector of modife option payoff from trajectory simulations.}

\item{p}{numeric value, power of the loss function, p > 0.}

\item{rate}{numeric value, risk free rate in the model, rate >= 0.}

\item{End_Time}{numeric value, end time of the option.}

\item{is_discounted}{logical value, if TRUE, payoff and modificate_payoff will be discounted.}
}
\description{
The Shortfall risk function takes payoff from option, and their modife version and returns a mean of shortfall weighted by x^p function.
}
\examples{
Shortfall_risk(5, 4, 1, 0.05, 1)
Shortfall_risk(c(100, 120), c(100, 107.5), 3, 0.05, TRUE)



}
\seealso{

}
