% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xi_put_price_explicit.R
\name{Xi_put_price_explicit}
\alias{Xi_put_price_explicit}
\title{Delta hedging for modified european put option}
\usage{
Xi_put_price_explicit(asset, strike, rate, vol, drift, p, time, End_Time, L, L2 = NA)
}
\arguments{
\item{asset}{a numeric vector of asset prices.}

\item{strike}{numeric value, strike price for call or put option.}

\item{rate}{numeric value, risk free rate in the model, r >= 0.}

\item{vol}{numeric value, volatility of the model, vol > 0.}

\item{drift}{numeric value, drift of the model.}

\item{p}{numeric value, power of the loss function, p > 1.}

\item{time}{a numeric vector of actual time, time > 0.}

\item{End_Time}{end time of the option, End_time >= time.}

\item{L2}{numeric value, determines option payoff, if L2 = NA, but is needed, function finds it with Newton's algorithm.}

\item{L}{numeric value, determines option payoff, see details, L > 0.}
}
\value{
A numeric vector, number of asset to hedge modification of european put option using x^p loss function.
}
\description{
The Xi_put_price_explicit function takes parameters from Black-Scholes model and returns a number of stock needed to fully hedge modified european put option.
}
\details{
Payoff of this modified put option is:
\subsection{\eqn{ 1(asset > L)(asset - strike)^+ }, when \eqn{ drift > rate }.}{
}

\subsection{\eqn{ 1(asset < L)(asset - strike)^+ }, when \eqn{ drift < rate }.}{
}

\subsection{\eqn{ L(asset - strike)^+ }, when \eqn{ drift == rate }, of course in this case L <= 1.}{
}
}
\examples{
Xi_put_price_explicit(100, 100, 0, 0.5, 0.05, 2,  0, 1, 105)
Xi_put_price_explicit(c(100, 120), 100, 0, 0.3, 0.05, 2, 0, 1, 105)
Xi_put_price_explicit(c(100, 120), 100, 0, 0.3, 0.05, 2, c(0, 0.5), 1, 105)



}
\seealso{
\url{https://en.wikipedia.org/wiki/Black–Scholes_model}.

}
