% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_payoff.R
\name{call_payoff}
\alias{call_payoff}
\title{Calculate european call option payoff}
\usage{
call_payoff(asset_price, strike)
}
\arguments{
\item{asset_price}{numeric vector, price of the asset over time.}

\item{strike}{numeric value, strike price of the option.}
}
\value{
A numeric value, european call option payoff.
}
\description{
The call_payoff function calculate payoffs from trajectory.
}
\examples{
x <- generate_scenarios(100, 0, 0.3)
call_payoff(x[, 1], strike = 100)
call_payoff(c(100, 120), strike = 107)
call_payoff(c(100, 120, 107), strike = 107)
}
